fix(uninstall): teardown-last execution order + 5 live-run bugs from LOG/Output
The first real Windows run surfaced a cluster of execution bugs. All fixed at
the root:

1. Teardown-last plan order (the user-prescribed sequence): tool binaries ->
   PATH -> "Shutdown (stopped last)" (daemon stop + broker service removal) ->
   data/cache/config (a running daemon holds handles inside them) -> "Runtime
   binaries (after shutdown)" (uffsd/uffs-broker/uffsmcp/uffs-mcp-http, whose
   images are locked while running) -> deferred self-delete of uffs.exe +
   uffs-update.exe. The working tools stay usable for the whole run.

2. One locked file no longer traps a whole directory: delete_binaries is now
   best-effort across the set (the original run lost 21 deletions to one
   lingering uffsd.exe), with a 750ms settle-and-retry pass, reporting exactly
   which files failed.

3. The daemon stop re-discovers the CURRENT daemon via the real
   `uffs --daemon kill` handler (pid file/socket) instead of the analyzed pid
   (stale after the deep sweep's coverage reload), then waits for IPC-down +
   image release before the runtime binaries are deleted.

4. Windows daemon-management elevation gate now mirrors the Unix owner gate:
   no elevation needed when there is no PID file (nothing to protect) or when
   the daemon's launch-state sidecar records a NON-elevated launch (the daemon
   now writes an "elevated" flag into daemon.state.json) — a user-level daemon
   is the user's to kill even with the broker gone. Falls back to the broker
   probe otherwise.

5. The deferred self-delete never actually deleted anything: std's Windows arg
   quoting backslash-escaped the `del "path"` quotes inside the `cmd /c`
   payload, which cmd.exe does not parse. Passed verbatim via raw_arg now.

6. Formatting: the coverage failure notes put "Continuing the deep sweep..."
   on its own line, and a blank separator precedes the `[sweep]` block (-v).

7. `uffs-broker --install` narrates its steps (sc create -> ok, sc start with
   a "can take a minute" note -> ok) instead of a silent minute-long wait.

plan.rs crossed the 800-LOC budget with the reorder; fixed by decomposing (the
established sibling-tests pattern): its unit tests moved to plan/tests.rs,
leaving plan.rs at 513 LOC. Tests updated to the new order contract plus a
regression test pinning the runtime-stem split. Validated with cargo clippy
(host) + cargo xwin clippy (Windows, prod + tests); 38 uninstall tests pass.
