+++ crates/uffs-broker/src/broker/service.rs
@@ -52,6 +52,21 @@ fn sc_output(output: &std::process::Output) -> String {
         .to_owned()
 }
 
+/// Print an in-progress step label without a trailing newline and flush, so
+/// the operator sees what a slow step (e.g. the blocking `sc start`) is doing
+/// before its "ok"/"failed" verdict lands on the same line.
+#[cfg(windows)]
+#[expect(
+    clippy::print_stdout,
+    reason = "CLI admin command — stdout is the user-visible result channel"
+)]
+fn print_step(label: &str) {
+    use std::io::Write as _;
+
+    print!("{label}");
+    let _flushed = std::io::stdout().flush();
+}
+
 /// Register the broker as an auto-start Windows Service and start it.
 ///
 /// # Why the argv is split the way it is
@@ -79,7 +94,11 @@ pub(super) fn install_service() -> anyhow::Result<()> {
         );
     }
 
+    // Step-by-step narration: `sc start` blocks until the service reports
+    // ready, which can take a minute — a silent wait reads as a hang.
     let exe = std::env::current_exe()?;
+    println!("Installing the UFFS Access Broker service...");
+    print_step("  registering the service (sc create)... ");
     let create = std::process::Command::new("sc.exe")
         .args([
             "create",
@@ -94,6 +113,7 @@ pub(super) fn install_service() -> anyhow::Result<()> {
         .output()?;
 
     if !create.status.success() {
+        println!("failed");
         // AUDIT-OK(bytes): `sc` output surfaced verbatim to the operator —
         // display only, no decision.
         anyhow::bail!(
@@ -102,21 +122,28 @@ pub(super) fn install_service() -> anyhow::Result<()> {
             sc_output(&create)
         );
     }
+    println!("ok");
 
     // Start it now so the broker is usable immediately — the whole point
     // is "no future UAC", which only holds once the service is running.
     // `start= auto` also brings it back on every boot.
+    print_step(
+        "  starting the service (Windows waits for it to report ready; \
+         this can take a minute)... ",
+    );
     let start = std::process::Command::new("sc.exe")
         .args(["start", SERVICE_NAME])
         .output()?;
 
     if start.status.success() {
+        println!("ok");
         println!(
             "UFFS Access Broker installed and started (auto-start on boot).\n\
              Non-elevated `uffs` searches will now use the broker for volume \
              access — no more UAC prompts."
         );
     } else {
+        println!("failed");
         // AUDIT-OK(bytes): `sc` output surfaced verbatim to the operator.
         println!(
             "Service installed (auto-start on boot), but starting it failed: \
