+++ crates/uffs-cli/src/commands/uninstall/plan.rs
@@ -226,68 +226,24 @@ pub(crate) fn build_plan(
 ) -> RemovalPlan {
     let mut groups: Vec<PlanGroup> = Vec::new();
 
-    // 1. Services (the broker, elevated) — removed first conceptually.
-    if inventory.broker_service == BrokerServiceState::Installed {
-        let item = PlanItem {
-            target: PlanTarget::RemoveService {
-                service: uffs_broker_protocol::SERVICE_NAME.to_owned(),
-            },
-            needs_elevation: true,
-            scope: ItemScope::Machine,
-            bytes: 0,
-        };
-        push_group(&mut groups, "Services", vec![item], args.scope);
-    }
+    // The working tools stay alive until the very end: tool binaries first,
+    // then PATH, then the shutdown of the running parts (daemon process +
+    // broker service), then the data dirs they had open, and finally the
+    // runtime binaries whose images were locked until that shutdown. The
+    // running uffs.exe / uffs-update.exe are deferred past process exit
+    // (self-delete) by the executor.
 
-    // 2. Processes (stopped before their binaries are deleted). The broker is a
-    // LocalSystem **service** — `taskkill` can't stop it (returns exit 128, and
-    // the SCM would just restart it), so it is never a StopProcess item; the
-    // RemoveService item above stops + deletes it via `sc`. The daemon / MCP are
-    // ordinary user-owned processes, so a plain stop applies and needs no admin.
-    let processes: Vec<PlanItem> = report
-        .running
+    // 1. Tool binaries — per root: unmanaged/dev delete, winget delegate. The
+    // runtime binaries (daemon, broker, MCP servers) are split into the final
+    // group below: their images are locked while those processes/services run.
+    let binaries: Vec<PlanItem> = report
+        .roots
         .iter()
-        .filter(|process| !matches!(process.component, Component::Broker))
-        .map(|process| PlanItem {
-            target: PlanTarget::StopProcess {
-                component: process.component.label().to_owned(),
-                pid: process.pid,
-            },
-            needs_elevation: false,
-            scope: ItemScope::Any,
-            bytes: 0,
-        })
+        .filter_map(|root| binary_item(root, StemSet::Tools))
         .collect();
-    push_group(
-        &mut groups,
-        "Processes (stopped first)",
-        processes,
-        args.scope,
-    );
-
-    // 3. Binaries — per root: unmanaged/dev delete, winget delegate.
-    let binaries: Vec<PlanItem> = report.roots.iter().filter_map(binary_item).collect();
     push_group(&mut groups, "Binaries", binaries, args.scope);
 
-    // 4. Data / cache / config dirs that exist (skip config under --keep-config).
-    let dirs: Vec<PlanItem> = inventory
-        .dirs
-        .iter()
-        .filter(|dir| dir.exists)
-        .filter(|dir| !(args.keep_config && dir.kind == ArtifactKind::Config))
-        .map(|dir| PlanItem {
-            target: PlanTarget::DeleteDir {
-                path: dir.path.clone(),
-                label: dir.kind.label(),
-            },
-            needs_elevation: false,
-            scope: ItemScope::User,
-            bytes: dir.size_bytes,
-        })
-        .collect();
-    push_group(&mut groups, "Data / cache / config", dirs, args.scope);
-
-    // 5. PATH entries that point at a removed unmanaged/dev root that is
+    // 2. PATH entries that point at a removed unmanaged/dev root that is
     // *dedicated* to UFFS (only uffs* files) — provably ours, so safe to drop. A
     // shared bin dir (~/bin, ~/.local/bin) is filtered out upstream and never
     // appears here. WinGet roots are managed by winget. Skipped under --no-path.
@@ -319,6 +275,74 @@ pub(crate) fn build_plan(
             .collect();
         push_group(&mut groups, "PATH", path_items, args.scope);
     }
+
+    // 3. Shutdown of the running parts — LAST among the live pieces, so the
+    // tooling stays usable during the run. The broker is a LocalSystem
+    // **service** — `taskkill` can't stop it (returns exit 128, and the SCM
+    // would just restart it), so it is never a StopProcess item; the
+    // RemoveService item stops + deletes it via `sc`. The daemon / MCP are
+    // ordinary user-owned processes, so a plain stop applies and needs no
+    // admin. (At execution the daemon is re-discovered by its pid file — the
+    // analyzed pid can go stale when the deep sweep reloads it.)
+    let mut shutdown: Vec<PlanItem> = report
+        .running
+        .iter()
+        .filter(|process| !matches!(process.component, Component::Broker))
+        .map(|process| PlanItem {
+            target: PlanTarget::StopProcess {
+                component: process.component.label().to_owned(),
+                pid: process.pid,
+            },
+            needs_elevation: false,
+            scope: ItemScope::Any,
+            bytes: 0,
+        })
+        .collect();
+    if inventory.broker_service == BrokerServiceState::Installed {
+        shutdown.push(PlanItem {
+            target: PlanTarget::RemoveService {
+                service: uffs_broker_protocol::SERVICE_NAME.to_owned(),
+            },
+            needs_elevation: true,
+            scope: ItemScope::Machine,
+            bytes: 0,
+        });
+    }
+    push_group(&mut groups, "Shutdown (stopped last)", shutdown, args.scope);
+
+    // 4. Data / cache / config dirs that exist (skip config under
+    // --keep-config). After the daemon shutdown: a running daemon holds open
+    // handles (pid file, socket, mmap'd caches) inside these dirs.
+    let dirs: Vec<PlanItem> = inventory
+        .dirs
+        .iter()
+        .filter(|dir| dir.exists)
+        .filter(|dir| !(args.keep_config && dir.kind == ArtifactKind::Config))
+        .map(|dir| PlanItem {
+            target: PlanTarget::DeleteDir {
+                path: dir.path.clone(),
+                label: dir.kind.label(),
+            },
+            needs_elevation: false,
+            scope: ItemScope::User,
+            bytes: dir.size_bytes,
+        })
+        .collect();
+    push_group(&mut groups, "Data / cache / config", dirs, args.scope);
+
+    // 5. Runtime binaries — deletable only now that their processes/services
+    // are stopped (Windows locks a running image).
+    let runtime: Vec<PlanItem> = report
+        .roots
+        .iter()
+        .filter_map(|root| binary_item(root, StemSet::Runtime))
+        .collect();
+    push_group(
+        &mut groups,
+        "Runtime binaries (after shutdown)",
+        runtime,
+        args.scope,
+    );
 
     RemovalPlan { groups }
 }
@@ -361,8 +385,32 @@ fn paths_equal_ignore_case(left: &Path, right: &Path) -> bool {
         .eq_ignore_ascii_case(&right.to_string_lossy())
 }
 
-/// Build the per-root binary plan item, or `None` for an empty root.
-fn binary_item(root: &InstallRoot) -> Option<PlanItem> {
+/// Binary stems whose images are locked while the resident parts run (the
+/// daemon, the broker service, the MCP servers). Deleted in the final plan
+/// group, after the shutdown items; every other stem is a plain tool binary.
+const RUNTIME_STEMS: &[&str] = &["uffsd", "uffs-broker", "uffsmcp", "uffs-mcp-http"];
+
+/// Which slice of a root's binaries a [`binary_item`] call covers.
+#[derive(Clone, Copy, PartialEq, Eq)]
+enum StemSet {
+    /// Plain tool binaries — deletable any time (group 1).
+    Tools,
+    /// [`RUNTIME_STEMS`] — deletable only after the shutdown group.
+    Runtime,
+}
+
+/// Whether `stem` names a runtime binary (see [`RUNTIME_STEMS`]).
+fn is_runtime_stem(stem: &str) -> bool {
+    RUNTIME_STEMS
+        .iter()
+        .any(|runtime| runtime.eq_ignore_ascii_case(stem))
+}
+
+/// Build the per-root binary plan item for the requested stem set, or `None`
+/// when the root has no matching binaries. A `WinGet` root delegates whole to
+/// `winget uninstall` in the Tools pass (winget owns the stop/delete order for
+/// its own package), so its Runtime pass is empty.
+fn binary_item(root: &InstallRoot, set: StemSet) -> Option<PlanItem> {
     if root.binaries.is_empty() {
         return None;
     }
@@ -373,15 +421,31 @@ fn binary_item(root: &InstallRoot) -> Option<PlanItem> {
         ItemScope::User
     };
     let target = match root.channel {
-        Channel::WinGet => PlanTarget::DelegateWinget {
-            package_id: WINGET_PACKAGE_ID.to_owned(),
-            scope: root.scope,
-            dir: root.dir.clone(),
-        },
-        Channel::Unmanaged | Channel::DevBuild | Channel::Unknown => PlanTarget::DeleteBinaries {
-            dir: root.dir.clone(),
-            stems: root.binaries.iter().map(|bin| bin.name.clone()).collect(),
-        },
+        Channel::WinGet => {
+            if set == StemSet::Runtime {
+                return None;
+            }
+            PlanTarget::DelegateWinget {
+                package_id: WINGET_PACKAGE_ID.to_owned(),
+                scope: root.scope,
+                dir: root.dir.clone(),
+            }
+        }
+        Channel::Unmanaged | Channel::DevBuild | Channel::Unknown => {
+            let stems: Vec<String> = root
+                .binaries
+                .iter()
+                .filter(|bin| (set == StemSet::Runtime) == is_runtime_stem(&bin.name))
+                .map(|bin| bin.name.clone())
+                .collect();
+            if stems.is_empty() {
+                return None;
+            }
+            PlanTarget::DeleteBinaries {
+                dir: root.dir.clone(),
+                stems,
+            }
+        }
     };
     Some(PlanItem {
         target,
@@ -446,336 +510,4 @@ const fn scope_admits(requested: UninstallScope, item: ItemScope) -> bool {
 }
 
 #[cfg(test)]
-mod tests {
-    use std::path::PathBuf;
-
-    #[cfg(windows)]
-    use super::build_stray_plan;
-    use super::{PlanTarget, RemovalPlan, build_plan};
-    use crate::commands::uninstall::args::{UninstallArgs, UninstallScope};
-    use crate::commands::uninstall::inventory::{
-        ArtifactDir, ArtifactKind, BrokerServiceState, Inventory,
-    };
-    use crate::commands::update::model::{
-        BinaryInfo, Channel, Component, DetectionReport, InstallRoot, RunningProcess, Scope,
-    };
-
-    fn root(channel: Channel, scope: Scope, dir: &str) -> InstallRoot {
-        InstallRoot {
-            dir: PathBuf::from(dir),
-            channel,
-            scope,
-            anchored_by: Vec::new(),
-            binaries: vec![BinaryInfo {
-                name: "uffs".to_owned(),
-                version: Some("0.6.16".to_owned()),
-            }],
-        }
-    }
-
-    fn inventory(broker: BrokerServiceState, config_size: u64) -> Inventory {
-        Inventory {
-            dirs: vec![
-                ArtifactDir {
-                    kind: ArtifactKind::Cache,
-                    path: PathBuf::from("/x/cache"),
-                    exists: true,
-                    size_bytes: 2048,
-                },
-                ArtifactDir {
-                    kind: ArtifactKind::Config,
-                    path: PathBuf::from("/x/config"),
-                    exists: true,
-                    size_bytes: config_size,
-                },
-            ],
-            broker_service: broker,
-        }
-    }
-
-    fn has_target(plan: &RemovalPlan, predicate: impl Fn(&PlanTarget) -> bool) -> bool {
-        plan.items().any(|item| predicate(&item.target))
-    }
-
-    /// Build a plan with no PATH entries (PATH has its own dedicated test).
-    fn built(report: &DetectionReport, inventory: &Inventory, args: &UninstallArgs) -> RemovalPlan {
-        build_plan(report, inventory, args, &[])
-    }
-
-    #[test]
-    fn winget_root_is_delegated_not_deleted() {
-        let report = DetectionReport {
-            roots: vec![root(Channel::WinGet, Scope::User, r"C:\winget\uffs")],
-            running: Vec::new(),
-        };
-        let plan = built(
-            &report,
-            &inventory(BrokerServiceState::Absent, 1024),
-            &UninstallArgs::default(),
-        );
-        assert!(has_target(&plan, |target| matches!(
-            target,
-            PlanTarget::DelegateWinget { .. }
-        )));
-        assert!(!has_target(&plan, |target| matches!(
-            target,
-            PlanTarget::DeleteBinaries { .. }
-        )));
-    }
-
-    #[test]
-    fn machine_root_needs_elevation() {
-        let report = DetectionReport {
-            roots: vec![root(
-                Channel::Unmanaged,
-                Scope::Machine,
-                r"C:\Program Files\uffs",
-            )],
-            running: Vec::new(),
-        };
-        let plan = built(
-            &report,
-            &inventory(BrokerServiceState::Absent, 1024),
-            &UninstallArgs::default(),
-        );
-        assert!(plan.requires_elevation());
-    }
-
-    #[test]
-    fn service_present_requires_elevation_and_is_first() {
-        let report = DetectionReport {
-            roots: Vec::new(),
-            running: Vec::new(),
-        };
-        let plan = built(
-            &report,
-            &inventory(BrokerServiceState::Installed, 1024),
-            &UninstallArgs::default(),
-        );
-        assert!(plan.requires_elevation());
-        assert!(has_target(&plan, |target| matches!(
-            target,
-            PlanTarget::RemoveService { .. }
-        )));
-        assert_eq!(plan.groups.first().expect("a group").title, "Services");
-    }
-
-    #[test]
-    fn keep_config_drops_the_config_dir() {
-        let report = DetectionReport {
-            roots: Vec::new(),
-            running: Vec::new(),
-        };
-        let inv = inventory(BrokerServiceState::Absent, 4096);
-        let with_config = built(&report, &inv, &UninstallArgs::default());
-        let keep = UninstallArgs {
-            keep_config: true,
-            ..UninstallArgs::default()
-        };
-        let without_config = built(&report, &inv, &keep);
-        assert!(with_config.total_bytes() > without_config.total_bytes());
-    }
-
-    #[test]
-    fn scope_user_excludes_the_machine_service() {
-        let report = DetectionReport {
-            roots: Vec::new(),
-            running: Vec::new(),
-        };
-        let user_only = UninstallArgs {
-            scope: UninstallScope::User,
-            ..UninstallArgs::default()
-        };
-        let plan = built(
-            &report,
-            &inventory(BrokerServiceState::Installed, 1024),
-            &user_only,
-        );
-        assert!(!has_target(&plan, |target| matches!(
-            target,
-            PlanTarget::RemoveService { .. }
-        )));
-        assert!(!plan.requires_elevation());
-    }
-
-    #[test]
-    fn running_process_becomes_a_stop_item() {
-        let report = DetectionReport {
-            roots: Vec::new(),
-            running: vec![RunningProcess {
-                component: Component::Daemon,
-                pid: 4242,
-                image_path: None,
-                command_line: None,
-                version: None,
-            }],
-        };
-        let plan = built(
-            &report,
-            &inventory(BrokerServiceState::Absent, 1024),
-            &UninstallArgs::default(),
-        );
-        assert!(has_target(&plan, |target| matches!(
-            target,
-            PlanTarget::StopProcess { .. }
-        )));
-    }
-
-    #[test]
-    fn drop_elevation_required_removes_broker_keeps_the_rest() {
-        let report = DetectionReport {
-            roots: Vec::new(),
-            running: vec![
-                RunningProcess {
-                    component: Component::Broker,
-                    pid: 11,
-                    image_path: None,
-                    command_line: None,
-                    version: None,
-                },
-                RunningProcess {
-                    component: Component::Daemon,
-                    pid: 22,
-                    image_path: None,
-                    command_line: None,
-                    version: None,
-                },
-            ],
-        };
-        // Broker service installed -> an admin-only RemoveService item. The
-        // broker *process* is filtered out (it's a service, stopped via sc, not
-        // taskkill); only the user-owned daemon stop remains, needing no admin.
-        let mut plan = built(
-            &report,
-            &inventory(BrokerServiceState::Installed, 1024),
-            &UninstallArgs::default(),
-        );
-        assert!(
-            plan.requires_elevation(),
-            "broker service + process need admin"
-        );
-
-        let dropped = plan.drop_elevation_required();
-        assert!(!plan.requires_elevation(), "admin-only items were dropped");
-        assert!(
-            !dropped.is_empty() && dropped.iter().all(|desc| !desc.is_empty()),
-            "the dropped items are returned as human descriptions for the summary"
-        );
-        assert!(
-            !has_target(&plan, |target| matches!(
-                target,
-                PlanTarget::RemoveService { .. }
-            )),
-            "the broker service item is gone"
-        );
-        let stop_pids: Vec<u32> = plan
-            .items()
-            .filter_map(|item| {
-                if let PlanTarget::StopProcess { pid, .. } = &item.target {
-                    Some(*pid)
-                } else {
-                    None
-                }
-            })
-            .collect();
-        assert_eq!(stop_pids, vec![22], "only the daemon stop survives");
-    }
-
-    #[test]
-    #[cfg(windows)]
-    fn stray_plan_is_one_group_of_unprivileged_delete_file_items() {
-        use crate::commands::uninstall::sweep::StrayHit;
-
-        assert!(build_stray_plan(&[]).is_empty(), "no strays -> empty plan");
-        let strays = vec![
-            StrayHit {
-                path: PathBuf::from("/home/me/Downloads/uffs"),
-                version: Some("0.5.0".to_owned()),
-            },
-            StrayHit {
-                path: PathBuf::from("/tmp/x_compact.uffs"),
-                version: None,
-            },
-        ];
-        let plan = build_stray_plan(&strays);
-        assert_eq!(plan.item_count(), 2);
-        assert!(
-            plan.items()
-                .all(|item| matches!(item.target, PlanTarget::DeleteFile { .. })),
-            "every stray item is a DeleteFile"
-        );
-        assert!(
-            !plan.requires_elevation(),
-            "strays never require up-front elevation (best-effort on failure)"
-        );
-    }
-
-    #[test]
-    fn path_entry_matching_a_removed_root_is_offered_and_respects_no_path() {
-        let report = DetectionReport {
-            roots: vec![root(Channel::Unmanaged, Scope::User, r"C:\Users\me\bin")],
-            running: Vec::new(),
-        };
-        let inv = inventory(BrokerServiceState::Absent, 1024);
-        // The 4th arg is the already-vetted removable-dir set; a case-insensitive
-        // match to the removed root → offered. (Exclusivity vetting is tested in
-        // analyze::removable_path_dirs; here we exercise build_plan's emission.)
-        let on_path = [PathBuf::from(r"c:\users\me\bin")];
-        let offered = build_plan(&report, &inv, &UninstallArgs::default(), &on_path);
-        assert!(has_target(&offered, |target| matches!(
-            target,
-            PlanTarget::RemovePathEntry { .. }
-        )));
-        // --no-path suppresses the PATH group entirely.
-        let no_path = UninstallArgs {
-            no_path: true,
-            ..UninstallArgs::default()
-        };
-        let suppressed = build_plan(&report, &inv, &no_path, &on_path);
-        assert!(!has_target(&suppressed, |target| matches!(
-            target,
-            PlanTarget::RemovePathEntry { .. }
-        )));
-        // A PATH entry that does not match any root is never touched.
-        let unrelated = [PathBuf::from(r"C:\unrelated")];
-        let untouched = build_plan(&report, &inv, &UninstallArgs::default(), &unrelated);
-        assert!(!has_target(&untouched, |target| matches!(
-            target,
-            PlanTarget::RemovePathEntry { .. }
-        )));
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn unix_user_writable_root_skips_escalation_root_owned_flags_it() {
-        use std::path::Path;
-
-        use super::binaries_need_escalation;
-        // The temp dir is user-writable → removable without sudo.
-        assert!(!binaries_need_escalation(
-            Scope::Unknown,
-            &std::env::temp_dir()
-        ));
-        // A non-existent / unwritable path → flagged for escalation.
-        assert!(binaries_need_escalation(
-            Scope::Unknown,
-            Path::new("/nonexistent/uffs-escalation-probe")
-        ));
-    }
-
-    #[cfg(windows)]
-    #[test]
-    fn windows_escalation_follows_machine_scope() {
-        use std::path::Path;
-
-        use super::binaries_need_escalation;
-        assert!(binaries_need_escalation(
-            Scope::Machine,
-            Path::new(r"C:\Program Files\uffs")
-        ));
-        assert!(!binaries_need_escalation(
-            Scope::User,
-            Path::new(r"C:\Users\me\bin")
-        ));
-    }
-}
+mod tests;
