+++ crates/uffs-cli/src/commands/uninstall/plan/tests.rs
@@ -1,0 +1,406 @@
+// SPDX-License-Identifier: MPL-2.0
+// Copyright (c) 2025-2026 SKY, LLC.
+
+//! Unit tests for the removal-plan construction ([`super`]) — extracted into a
+//! sibling module (the `compact_cache/tests.rs` / `backend_tests.rs` pattern)
+//! so `plan.rs` stays within the file-size policy.
+
+use std::path::PathBuf;
+
+#[cfg(windows)]
+use super::build_stray_plan;
+use super::{PlanTarget, RemovalPlan, build_plan};
+use crate::commands::uninstall::args::{UninstallArgs, UninstallScope};
+use crate::commands::uninstall::inventory::{
+    ArtifactDir, ArtifactKind, BrokerServiceState, Inventory,
+};
+use crate::commands::update::model::{
+    BinaryInfo, Channel, Component, DetectionReport, InstallRoot, RunningProcess, Scope,
+};
+
+fn root(channel: Channel, scope: Scope, dir: &str) -> InstallRoot {
+    InstallRoot {
+        dir: PathBuf::from(dir),
+        channel,
+        scope,
+        anchored_by: Vec::new(),
+        binaries: vec![BinaryInfo {
+            name: "uffs".to_owned(),
+            version: Some("0.6.16".to_owned()),
+        }],
+    }
+}
+
+fn inventory(broker: BrokerServiceState, config_size: u64) -> Inventory {
+    Inventory {
+        dirs: vec![
+            ArtifactDir {
+                kind: ArtifactKind::Cache,
+                path: PathBuf::from("/x/cache"),
+                exists: true,
+                size_bytes: 2048,
+            },
+            ArtifactDir {
+                kind: ArtifactKind::Config,
+                path: PathBuf::from("/x/config"),
+                exists: true,
+                size_bytes: config_size,
+            },
+        ],
+        broker_service: broker,
+    }
+}
+
+fn has_target(plan: &RemovalPlan, predicate: impl Fn(&PlanTarget) -> bool) -> bool {
+    plan.items().any(|item| predicate(&item.target))
+}
+
+/// Build a plan with no PATH entries (PATH has its own dedicated test).
+fn built(report: &DetectionReport, inventory: &Inventory, args: &UninstallArgs) -> RemovalPlan {
+    build_plan(report, inventory, args, &[])
+}
+
+#[test]
+fn winget_root_is_delegated_not_deleted() {
+    let report = DetectionReport {
+        roots: vec![root(Channel::WinGet, Scope::User, r"C:\winget\uffs")],
+        running: Vec::new(),
+    };
+    let plan = built(
+        &report,
+        &inventory(BrokerServiceState::Absent, 1024),
+        &UninstallArgs::default(),
+    );
+    assert!(has_target(&plan, |target| matches!(
+        target,
+        PlanTarget::DelegateWinget { .. }
+    )));
+    assert!(!has_target(&plan, |target| matches!(
+        target,
+        PlanTarget::DeleteBinaries { .. }
+    )));
+}
+
+#[test]
+fn machine_root_needs_elevation() {
+    let report = DetectionReport {
+        roots: vec![root(
+            Channel::Unmanaged,
+            Scope::Machine,
+            r"C:\Program Files\uffs",
+        )],
+        running: Vec::new(),
+    };
+    let plan = built(
+        &report,
+        &inventory(BrokerServiceState::Absent, 1024),
+        &UninstallArgs::default(),
+    );
+    assert!(plan.requires_elevation());
+}
+
+#[test]
+fn service_present_requires_elevation_and_shuts_down_before_data() {
+    let report = DetectionReport {
+        roots: Vec::new(),
+        running: Vec::new(),
+    };
+    let plan = built(
+        &report,
+        &inventory(BrokerServiceState::Installed, 1024),
+        &UninstallArgs::default(),
+    );
+    assert!(plan.requires_elevation());
+    assert!(has_target(&plan, |target| matches!(
+        target,
+        PlanTarget::RemoveService { .. }
+    )));
+    // Teardown-last ordering: the tools stay usable during the run, so the
+    // shutdown group comes late — but still BEFORE the data dirs (a
+    // running daemon holds open handles inside them).
+    let titles: Vec<&str> = plan.groups.iter().map(|group| group.title).collect();
+    let shutdown = titles
+        .iter()
+        .position(|title| *title == "Shutdown (stopped last)")
+        .expect("a shutdown group");
+    let data = titles
+        .iter()
+        .position(|title| *title == "Data / cache / config")
+        .expect("a data group");
+    assert!(shutdown < data, "shutdown must precede data: {titles:?}");
+}
+
+#[test]
+fn runtime_binaries_split_into_the_post_shutdown_group() {
+    // A root holding both tool and runtime binaries: the tools delete in
+    // the first group; uffsd/uffs-broker (image locked while running) land
+    // in "Runtime binaries (after shutdown)", after the shutdown group.
+    let mut mixed = root(Channel::Unmanaged, Scope::User, "/opt/uffs");
+    mixed.binaries = ["uffs", "analyze-diff", "uffsd", "uffs-broker"]
+        .into_iter()
+        .map(|name| BinaryInfo {
+            name: name.to_owned(),
+            version: None,
+        })
+        .collect();
+    let report = DetectionReport {
+        roots: vec![mixed],
+        running: Vec::new(),
+    };
+    let plan = built(
+        &report,
+        &inventory(BrokerServiceState::Absent, 1024),
+        &UninstallArgs::default(),
+    );
+
+    let stems_of = |title: &str| -> Vec<String> {
+        plan.groups
+            .iter()
+            .find(|group| group.title == title)
+            .into_iter()
+            .flat_map(|group| &group.items)
+            .filter_map(|item| {
+                if let PlanTarget::DeleteBinaries { stems, .. } = &item.target {
+                    Some(stems.clone())
+                } else {
+                    None
+                }
+            })
+            .flatten()
+            .collect()
+    };
+    assert_eq!(stems_of("Binaries"), vec!["uffs", "analyze-diff"]);
+    assert_eq!(stems_of("Runtime binaries (after shutdown)"), vec![
+        "uffsd",
+        "uffs-broker"
+    ]);
+    let titles: Vec<&str> = plan.groups.iter().map(|group| group.title).collect();
+    let tools = titles
+        .iter()
+        .position(|title| *title == "Binaries")
+        .expect("tools");
+    let runtime = titles
+        .iter()
+        .position(|title| *title == "Runtime binaries (after shutdown)")
+        .expect("runtime");
+    assert!(tools < runtime, "runtime group must be last: {titles:?}");
+}
+
+#[test]
+fn keep_config_drops_the_config_dir() {
+    let report = DetectionReport {
+        roots: Vec::new(),
+        running: Vec::new(),
+    };
+    let inv = inventory(BrokerServiceState::Absent, 4096);
+    let with_config = built(&report, &inv, &UninstallArgs::default());
+    let keep = UninstallArgs {
+        keep_config: true,
+        ..UninstallArgs::default()
+    };
+    let without_config = built(&report, &inv, &keep);
+    assert!(with_config.total_bytes() > without_config.total_bytes());
+}
+
+#[test]
+fn scope_user_excludes_the_machine_service() {
+    let report = DetectionReport {
+        roots: Vec::new(),
+        running: Vec::new(),
+    };
+    let user_only = UninstallArgs {
+        scope: UninstallScope::User,
+        ..UninstallArgs::default()
+    };
+    let plan = built(
+        &report,
+        &inventory(BrokerServiceState::Installed, 1024),
+        &user_only,
+    );
+    assert!(!has_target(&plan, |target| matches!(
+        target,
+        PlanTarget::RemoveService { .. }
+    )));
+    assert!(!plan.requires_elevation());
+}
+
+#[test]
+fn running_process_becomes_a_stop_item() {
+    let report = DetectionReport {
+        roots: Vec::new(),
+        running: vec![RunningProcess {
+            component: Component::Daemon,
+            pid: 4242,
+            image_path: None,
+            command_line: None,
+            version: None,
+        }],
+    };
+    let plan = built(
+        &report,
+        &inventory(BrokerServiceState::Absent, 1024),
+        &UninstallArgs::default(),
+    );
+    assert!(has_target(&plan, |target| matches!(
+        target,
+        PlanTarget::StopProcess { .. }
+    )));
+}
+
+#[test]
+fn drop_elevation_required_removes_broker_keeps_the_rest() {
+    let report = DetectionReport {
+        roots: Vec::new(),
+        running: vec![
+            RunningProcess {
+                component: Component::Broker,
+                pid: 11,
+                image_path: None,
+                command_line: None,
+                version: None,
+            },
+            RunningProcess {
+                component: Component::Daemon,
+                pid: 22,
+                image_path: None,
+                command_line: None,
+                version: None,
+            },
+        ],
+    };
+    // Broker service installed -> an admin-only RemoveService item. The
+    // broker *process* is filtered out (it's a service, stopped via sc, not
+    // taskkill); only the user-owned daemon stop remains, needing no admin.
+    let mut plan = built(
+        &report,
+        &inventory(BrokerServiceState::Installed, 1024),
+        &UninstallArgs::default(),
+    );
+    assert!(
+        plan.requires_elevation(),
+        "broker service + process need admin"
+    );
+
+    let dropped = plan.drop_elevation_required();
+    assert!(!plan.requires_elevation(), "admin-only items were dropped");
+    assert!(
+        !dropped.is_empty() && dropped.iter().all(|desc| !desc.is_empty()),
+        "the dropped items are returned as human descriptions for the summary"
+    );
+    assert!(
+        !has_target(&plan, |target| matches!(
+            target,
+            PlanTarget::RemoveService { .. }
+        )),
+        "the broker service item is gone"
+    );
+    let stop_pids: Vec<u32> = plan
+        .items()
+        .filter_map(|item| {
+            if let PlanTarget::StopProcess { pid, .. } = &item.target {
+                Some(*pid)
+            } else {
+                None
+            }
+        })
+        .collect();
+    assert_eq!(stop_pids, vec![22], "only the daemon stop survives");
+}
+
+#[test]
+#[cfg(windows)]
+fn stray_plan_is_one_group_of_unprivileged_delete_file_items() {
+    use crate::commands::uninstall::sweep::StrayHit;
+
+    assert!(build_stray_plan(&[]).is_empty(), "no strays -> empty plan");
+    let strays = vec![
+        StrayHit {
+            path: PathBuf::from("/home/me/Downloads/uffs"),
+            version: Some("0.5.0".to_owned()),
+        },
+        StrayHit {
+            path: PathBuf::from("/tmp/x_compact.uffs"),
+            version: None,
+        },
+    ];
+    let plan = build_stray_plan(&strays);
+    assert_eq!(plan.item_count(), 2);
+    assert!(
+        plan.items()
+            .all(|item| matches!(item.target, PlanTarget::DeleteFile { .. })),
+        "every stray item is a DeleteFile"
+    );
+    assert!(
+        !plan.requires_elevation(),
+        "strays never require up-front elevation (best-effort on failure)"
+    );
+}
+
+#[test]
+fn path_entry_matching_a_removed_root_is_offered_and_respects_no_path() {
+    let report = DetectionReport {
+        roots: vec![root(Channel::Unmanaged, Scope::User, r"C:\Users\me\bin")],
+        running: Vec::new(),
+    };
+    let inv = inventory(BrokerServiceState::Absent, 1024);
+    // The 4th arg is the already-vetted removable-dir set; a case-insensitive
+    // match to the removed root → offered. (Exclusivity vetting is tested in
+    // analyze::removable_path_dirs; here we exercise build_plan's emission.)
+    let on_path = [PathBuf::from(r"c:\users\me\bin")];
+    let offered = build_plan(&report, &inv, &UninstallArgs::default(), &on_path);
+    assert!(has_target(&offered, |target| matches!(
+        target,
+        PlanTarget::RemovePathEntry { .. }
+    )));
+    // --no-path suppresses the PATH group entirely.
+    let no_path = UninstallArgs {
+        no_path: true,
+        ..UninstallArgs::default()
+    };
+    let suppressed = build_plan(&report, &inv, &no_path, &on_path);
+    assert!(!has_target(&suppressed, |target| matches!(
+        target,
+        PlanTarget::RemovePathEntry { .. }
+    )));
+    // A PATH entry that does not match any root is never touched.
+    let unrelated = [PathBuf::from(r"C:\unrelated")];
+    let untouched = build_plan(&report, &inv, &UninstallArgs::default(), &unrelated);
+    assert!(!has_target(&untouched, |target| matches!(
+        target,
+        PlanTarget::RemovePathEntry { .. }
+    )));
+}
+
+#[cfg(unix)]
+#[test]
+fn unix_user_writable_root_skips_escalation_root_owned_flags_it() {
+    use std::path::Path;
+
+    use super::binaries_need_escalation;
+    // The temp dir is user-writable → removable without sudo.
+    assert!(!binaries_need_escalation(
+        Scope::Unknown,
+        &std::env::temp_dir()
+    ));
+    // A non-existent / unwritable path → flagged for escalation.
+    assert!(binaries_need_escalation(
+        Scope::Unknown,
+        Path::new("/nonexistent/uffs-escalation-probe")
+    ));
+}
+
+#[cfg(windows)]
+#[test]
+fn windows_escalation_follows_machine_scope() {
+    use std::path::Path;
+
+    use super::binaries_need_escalation;
+    assert!(binaries_need_escalation(
+        Scope::Machine,
+        Path::new(r"C:\Program Files\uffs")
+    ));
+    assert!(!binaries_need_escalation(
+        Scope::User,
+        Path::new(r"C:\Users\me\bin")
+    ));
+}
